Add Python 3 Compatibility
Could use a lot of cleanup but at this point twisted while i did patch
the parts we need to work its not officially fully ported to python 3.
Therefore i cannot completely test anyways. This is just for refference
later once we are done so we can just reapply this and move forward.