[UI|Daemon][#3081] Move baseargparser out of deluge/ui
This file is being used by both UI and Daemon. Therefor, it should not
be located in the UI, but in the top-level `deluge` and renamed to
`argparser_base`.
The same is with the function `set_dummy_trans`, which was located in
`translation_util` and is now part of `argparser_base`.