[Core] Changed on_alert_tracker_error behavior for lt >= 1.2
libtorrent 1.2 added endpoint struct to each tracker, to prevent false updates
we will need to verify that at least one endpoint to the errored tracker is working.
if there is at least one working, it will not set the tracker status to error and set it to
`Announce OK`. otherwise, it will use the error message from the alert.