[Tracker] Add a status for all trackers in a torrent
As part of the changes needed for a more informative trackers tab in the
UIs, we need the info for each of the trackers in the torrent.
For this, new status keys were added, and `tracker_status` is now
considered deprecated.

For tracker info, the keys are:
`trackers_status` - a dict of tracker_url->{status, message}
`trackers_peers` - a dict of tracker_url->number of peers in the tracker

`trackers_status` will contain the status of the torrent and also the
message from the relevant alert.